Initial index page
index.html is the default page that is loaded when you navigate to a
site.  So our first page should also be index.html.

A webpage should always have a doctype, and <html> tags
(both opening and closing).

Everything else goes inside the html tag.  Since it has no parent,
it is called the root element.

For more information:
https://developer.mozilla.org/en-US/docs/Web/HTML/Element/html