Added a title and mime-type
Inside the <head> tag, we put information that we want the browser to
see.

Using <title> we tell the browser what label to put in the browser tab.

Using <meta> we tell the browser (and any other programs that may read
our file) that this is a document with mime-type of HTML.  This tells
them to process this file as an HTML document

UTF-8 is a standard known as Unicode, which we can use to display any
character, from emojis to Chinese characters.