Added a page title
Notice how this title is in the <body>.
This means that this will be shown to the user.
Only content that is in the body will be shown to the user.

Notice how we are using <h1> tags to say that this is a big heading.
The <title> is NOT used here.  Only in the <head>

If you accidentally put this in the <head>, then it won't be displayed.