Added splash image of a white bull
Use <img> to add images.
* 'src' tells the browser from where to load the image.  Since the image
  is in the same location as the index.html file,
  'src="white_bull-1626x1080.jpg"' is all we need.
* ALWAYS include 'alt' text and make it descriptive English.  This is
  what a visually-impaired person's screen-reader will read out to them.
  This is also important for SEO and, if the image doesn't load, telling
  which image it was.
* 'title' is what will be shown in the tooltip that pops up when someone
  hovers over the image
* DO NOT include 'height' or 'width'.  This sort of styling should be
  done in your CSS

So that your webpage and its pictures load quickly, try to keep the
image file small.  Reduce the size of the picture to the maximum that
you will need on your website, save it, then compress the file.

I downloaded the image from Unsplash.com.
The photographer is https://unsplash.com/@samuelzeller

Notice that Unsplash has HD images.  Which means they are BIG!
The original size of the picture was 3549x2357 pixels, and the file
was over 3 megabytes (3MB = 3,000,000 bytes which is a lot.

Using the default picture editor on our PC, I reduced the size of the
image to 1626x1080 pixels (which is still HD and pretty big).  This size
is more suitable for desktops.  I would go even smaller if I was
targeting smaller screens, like tablets or phones.

Changing the size of the picture also reduces the size of the file.  But
then I compressed the file using https://tinyjpg.com/ and now the file
is just 200KB (2,000 bytes) and the image will work just as well as the
original one but load 100 times faster.