Added a heading and paragraph describing the page
For the second heading, I used an <h2> instead of an <h1>
The <h1> is the title of the article/content on the page.
Every following heading is thus a sub-heading and should have less
emphasis.  This hints to the reader that they are in a section of the
same article.

I liked the paragraph, so I pasted it into the README too.  I did it
in this commit as I didn't feel that this modification to the README was
a meaningful enough change to commit on its own.