Added some padding to the contents of <main>
Adding padding makes the size of <main> larger than the 80% that we set
it to.  This happens because the width and height is for the CONTENT excluding the border and padding

To change this, we change the box-sizing to border-box.
Now the border and padding and content, will be figured into the width
and height.
See https://developer.mozilla.org/en-US/docs/Web/CSS/box-sizing